backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
backports.tarfile==1.2.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-log==0.4.0
colorama==0.4.6
cryptography==45.0.5
deprecation==2.1.0
dnspython==2.7.0
docutils==0.21.2
dotty-dict==1.3.1
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
invoke==1.7.3
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
keyring==25.6.0
more-itertools==10.7.0
nh3==0.2.22
packaging==25.0
pkginfo==1.12.1.2
postgrest==1.1.1
postgrest-py==0.10.3
pyasn1==0.6.1
pycparser==2.22
pydantic==1.10.22
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-gitlab==3.15.0
python-jose==3.5.0
python-multipart==0.0.20
python-semantic-release==7.33.0
pywin32-ctypes==0.2.3
readme_renderer==44.0
realtime==2.5.3
requests==2.32.4
requests-toolbelt==1.0.0
rfc3986==1.5.0
rsa==4.9.1
semver==2.13.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.1
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
tomlkit==0.13.3
tqdm==4.67.1
twine==3.8.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==12.0
zipp==3.23.0
